refactor(web): extrae vistas de reservas y pasos de examen
COD-05: BookingsPage (1385->23) pasa a dispatcher e mueve Student/Tutor/
Teacher/AdminView y compartidos a pages/bookings/; ExamPage (1136->401)
extrae Config/InProgress/ResultsStep a pages/exam/ conservando el estado.
Render y comportamiento identicos.
