+++ apps/web/src/pages/ExamPage.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import { useSearchParams, useNavigate } from 'react-router-dom';
 import {
   useExamBankInfo,
@@ -7,749 +7,15 @@ import {
   useExamHistory,
   useStartAiAttempt,
 } from '../hooks/useExams';
-import { useMyCertificates } from '../hooks/useCertificates';
-import { downloadCertificatePdf } from '../utils/certificatePdf';
-import type { ExamAttemptStarted, ExamAttemptResult, ExamQuestionPublic } from '@vkbacademy/shared';
-import { downloadExamPdf } from '../utils/examPdf';
+import type { ExamAttemptStarted, ExamAttemptResult } from '@vkbacademy/shared';
 import { getApiErrorMessage } from '../utils/errorMessage';
+import { ConfigStep } from './exam/ConfigStep';
+import { InProgressStep } from './exam/InProgressStep';
+import { ResultsStep } from './exam/ResultsStep';
 
 // ─── Tipos internos ───────────────────────────────────────────────────────────
 
 type ExamState = 'config' | 'in-progress' | 'results';
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-function scoreColor(score: number): string {
-  if (score >= 70) return '#16a34a';
-  if (score >= 50) return '#ea580c';
-  return '#dc2626';
-}
-
-function scoreGradient(score: number): string {
-  if (score >= 70) return 'linear-gradient(135deg, #16a34a 0%, #22c55e 100%)';
-  if (score >= 50) return 'linear-gradient(135deg, #ea580c 0%, #f97316 100%)';
-  return 'linear-gradient(135deg, #dc2626 0%, #ef4444 100%)';
-}
-
-// ─── Componente: Configuración ────────────────────────────────────────────────
-
-function ConfigStep({
-  scopeTitle,
-  maxQuestions,
-  onStart,
-  isLoading,
-  recentAttempts,
-}: {
-  scopeTitle: string;
-  maxQuestions: number;
-  onStart: (numQ: number, timeLimit?: number, onlyOnce?: boolean) => void;
-  isLoading: boolean;
-  recentAttempts: { attemptId: string; score: number; numQuestions: number; submittedAt: string }[];
-}) {
-  const [numQ, setNumQ] = useState(Math.min(10, maxQuestions));
-  const [useTimer, setUseTimer] = useState(false);
-  const [timeMins, setTimeMins] = useState(15);
-  const [onlyOnce, setOnlyOnce] = useState(false);
-
-  const inputStyle: React.CSSProperties = {
-    padding: '10px 14px',
-    borderRadius: 'var(--radius-sm)',
-    border: '1.5px solid var(--color-border)',
-    background: 'var(--color-bg)',
-    color: 'var(--color-text)',
-    fontSize: '0.95rem',
-    outline: 'none',
-    transition: 'border-color 0.15s, box-shadow 0.15s',
-    boxSizing: 'border-box' as const,
-  };
-
-  const toggleRowStyle: React.CSSProperties = {
-    display: 'flex',
-    alignItems: 'center',
-    gap: 10,
-    padding: '12px 0',
-    borderBottom: '1px solid var(--color-border)',
-    cursor: 'pointer',
-    fontSize: '0.9rem',
-    color: 'var(--color-text)',
-  };
-
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column' as const, gap: 20 }}>
-      {/* Card de configuracion */}
-      <div className="vkb-card" style={{ padding: '28px' }}>
-        <h3
-          style={{
-            fontSize: '1rem',
-            fontWeight: 700,
-            color: 'var(--color-text)',
-            marginBottom: 20,
-          }}
-        >
-          {scopeTitle}
-        </h3>
-
-        {/* Numero de preguntas */}
-        <div style={{ marginBottom: 20 }}>
-          <label
-            style={{
-              display: 'block',
-              fontSize: '0.85rem',
-              fontWeight: 600,
-              color: 'var(--color-text-muted)',
-              marginBottom: 8,
-            }}
-          >
-            Numero de preguntas (maximo: {maxQuestions})
-          </label>
-          <div style={{ display: 'flex', alignItems: 'center', gap: 12 }}>
-            <input
-              type="number"
-              min={1}
-              max={maxQuestions}
-              value={numQ}
-              onChange={(e) => setNumQ(Math.min(maxQuestions, Math.max(1, Number(e.target.value))))}
-              style={{ ...inputStyle, width: 100 }}
-            />
-            <div className="progress-bar" style={{ flex: 1 }}>
-              <div className="progress-fill" style={{ width: `${(numQ / maxQuestions) * 100}%` }} />
-            </div>
-            <span
-              style={{
-                fontSize: '0.8rem',
-                color: 'var(--color-text-muted)',
-                fontWeight: 600,
-                whiteSpace: 'nowrap' as const,
-              }}
-            >
-              {numQ}/{maxQuestions}
-            </span>
-          </div>
-        </div>
-
-        {/* Toggle: limite de tiempo */}
-        <label style={toggleRowStyle}>
-          <input
-            id="use-timer"
-            type="checkbox"
-            checked={useTimer}
-            onChange={(e) => setUseTimer(e.target.checked)}
-            style={{ accentColor: 'var(--color-primary)', width: 16, height: 16 }}
-          />
-          <span style={{ fontWeight: 500 }}>Limite de tiempo</span>
-        </label>
-        {useTimer && (
-          <div
-            style={{ display: 'flex', alignItems: 'center', gap: 10, padding: '10px 0 12px 26px' }}
-          >
-            <input
-              type="number"
-              min={1}
-              max={180}
-              value={timeMins}
-              onChange={(e) => setTimeMins(Math.max(1, Number(e.target.value)))}
-              style={{ ...inputStyle, width: 80 }}
-            />
-            <span style={{ fontSize: '0.875rem', color: 'var(--color-text-muted)' }}>minutos</span>
-          </div>
-        )}
-
-        {/* Toggle: respuesta unica */}
-        <label style={{ ...toggleRowStyle, borderBottom: 'none', marginBottom: 0 }}>
-          <input
-            id="only-once"
-            type="checkbox"
-            checked={onlyOnce}
-            onChange={(e) => setOnlyOnce(e.target.checked)}
-            style={{ accentColor: 'var(--color-primary)', width: 16, height: 16 }}
-          />
-          <span style={{ fontWeight: 500 }}>Respuesta unica (no se puede cambiar)</span>
-        </label>
-
-        <button
-          className="btn btn-primary"
-          style={{ marginTop: 24, width: '100%', fontSize: '1rem', padding: '13px' }}
-          disabled={isLoading}
-          onClick={() => onStart(numQ, useTimer ? timeMins * 60 : undefined, onlyOnce || undefined)}
-        >
-          {isLoading ? 'Preparando...' : 'Comenzar examen'}
-        </button>
-      </div>
-
-      {/* Historial reciente */}
-      {recentAttempts.length > 0 && (
-        <div className="vkb-card" style={{ padding: '24px' }}>
-          <h4
-            style={{
-              fontSize: '0.9rem',
-              fontWeight: 700,
-              color: 'var(--color-text)',
-              marginBottom: 16,
-            }}
-          >
-            Intentos recientes
-          </h4>
-          <div style={{ display: 'flex', flexDirection: 'column' as const, gap: 0 }}>
-            {recentAttempts.map((a) => (
-              <div
-                key={a.attemptId}
-                style={{
-                  display: 'flex',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                  padding: '10px 0',
-                  borderBottom: '1px solid var(--color-border)',
-                  fontSize: '0.875rem',
-                  gap: 12,
-                }}
-              >
-                <span style={{ color: 'var(--color-text-muted)' }}>
-                  {new Date(a.submittedAt).toLocaleDateString('es-ES')}
-                </span>
-                <span style={{ color: 'var(--color-text-muted)' }}>{a.numQuestions} preguntas</span>
-                <span
-                  style={{
-                    fontWeight: 800,
-                    color: scoreColor(a.score),
-                    fontSize: '0.95rem',
-                  }}
-                >
-                  {a.score.toFixed(1)}%
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ─── Componente: Examen en progreso ───────────────────────────────────────────
-
-function InProgressStep({
-  attempt,
-  onSubmit,
-  isLoading,
-}: {
-  attempt: ExamAttemptStarted;
-  onSubmit: (answers: { questionId: string; answerIds: string[] }[]) => void;
-  isLoading: boolean;
-}) {
-  // Mapa pregunta → ids seleccionados. Para SINGLE/TRUE_FALSE el array tiene
-  // siempre 1 elemento; para MULTIPLE puede tener varios.
-  const [selectedAnswers, setSelectedAnswers] = useState<Record<string, string[]>>({});
-  const [secondsLeft, setSecondsLeft] = useState<number | null>(attempt.timeLimit ?? null);
-
-  const handleSubmit = useCallback(() => {
-    const answers = Object.entries(selectedAnswers)
-      .filter(([, ids]) => ids.length > 0)
-      .map(([questionId, answerIds]) => ({ questionId, answerIds }));
-    onSubmit(answers);
-  }, [selectedAnswers, onSubmit]);
-
-  // Cuenta atras
-  useEffect(() => {
-    if (secondsLeft === null) return;
-    if (secondsLeft <= 0) {
-      handleSubmit();
-      return;
-    }
-    const timer = setTimeout(() => setSecondsLeft((s) => (s !== null ? s - 1 : null)), 1000);
-    return () => clearTimeout(timer);
-  }, [secondsLeft, handleSubmit]);
-
-  const answeredCount = Object.values(selectedAnswers).filter((ids) => ids.length > 0).length;
-  const totalCount = attempt.questions.length;
-  const progressPct = (answeredCount / totalCount) * 100;
-
-  const isTimeCritical = secondsLeft !== null && secondsLeft < 60;
-
-  const formatTime = (secs: number) => {
-    const m = Math.floor(secs / 60);
-    const s = secs % 60;
-    return `${m}:${s.toString().padStart(2, '0')}`;
-  };
-
-  const toggleAnswer = (question: ExamQuestionPublic, answerId: string) => {
-    const isMultiple = question.type === 'MULTIPLE';
-    setSelectedAnswers((prev) => {
-      const current = prev[question.id] ?? [];
-      // Bloqueo onlyOnce: si ya hay selección y no es la misma respuesta en MULTIPLE, no permitir cambios.
-      if (attempt.onlyOnce && current.length > 0 && !isMultiple) return prev;
-
-      if (isMultiple) {
-        const exists = current.includes(answerId);
-        const next = exists ? current.filter((id) => id !== answerId) : [...current, answerId];
-        return { ...prev, [question.id]: next };
-      }
-      // SINGLE / TRUE_FALSE: reemplazo
-      return { ...prev, [question.id]: [answerId] };
-    });
-  };
-
-  return (
-    <div>
-      {/* Barra de estado */}
-      <div
-        style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          alignItems: 'center',
-          marginBottom: 14,
-          gap: 12,
-          flexWrap: 'wrap' as const,
-        }}
-      >
-        <span style={{ fontSize: '0.875rem', color: 'var(--color-text-muted)', fontWeight: 500 }}>
-          {answeredCount} de {totalCount} respondidas
-        </span>
-
-        {/* Temporizador */}
-        {secondsLeft !== null && (
-          <div
-            style={{
-              display: 'flex',
-              alignItems: 'center',
-              gap: 8,
-              padding: '8px 16px',
-              borderRadius: 'var(--radius-md)',
-              background: isTimeCritical ? 'rgba(220,38,38,0.12)' : 'var(--color-surface)',
-              border: isTimeCritical
-                ? '1.5px solid rgba(220,38,38,0.35)'
-                : '1.5px solid var(--color-border)',
-              animation: isTimeCritical ? 'pulse-glow 0.8s ease-in-out infinite' : 'none',
-            }}
-          >
-            <span style={{ fontSize: '1rem' }}>⏱</span>
-            <span
-              style={{
-                fontSize: '1.1rem',
-                fontWeight: 800,
-                color: isTimeCritical ? '#dc2626' : 'var(--color-text)',
-                letterSpacing: '0.05em',
-              }}
-            >
-              {formatTime(secondsLeft)}
-            </span>
-          </div>
-        )}
-      </div>
-
-      {/* Barra de progreso */}
-      <div
-        className="progress-bar"
-        style={{ marginBottom: 28 }}
-        role="progressbar"
-        aria-valuenow={Math.round(progressPct)}
-        aria-valuemin={0}
-        aria-valuemax={100}
-      >
-        <div className="progress-fill" style={{ width: `${progressPct}%` }} />
-      </div>
-
-      {/* Preguntas */}
-      {attempt.questions.map((q, idx) => {
-        const chosen = selectedAnswers[q.id] ?? [];
-        const isMultiple = q.type === 'MULTIPLE';
-        return (
-          <div key={q.id} className="vkb-card" style={{ padding: '22px 24px', marginBottom: 16 }}>
-            <div
-              style={{
-                fontWeight: 700,
-                marginBottom: 6,
-                color: 'var(--color-text)',
-                fontSize: '0.975rem',
-                lineHeight: 1.4,
-              }}
-            >
-              <span style={{ color: 'var(--color-primary)', marginRight: 8, fontWeight: 900 }}>
-                {idx + 1}.
-              </span>
-              {q.text}
-            </div>
-
-            {/* Hint para MULTIPLE */}
-            {isMultiple && (
-              <div
-                style={{
-                  fontSize: '0.72rem',
-                  fontWeight: 700,
-                  color: 'var(--color-primary)',
-                  marginBottom: 12,
-                  letterSpacing: '0.04em',
-                  textTransform: 'uppercase' as const,
-                }}
-              >
-                ☑ Selecciona todas las correctas
-              </div>
-            )}
-
-            <div
-              style={{ display: 'flex', flexDirection: 'column' as const, gap: 8, marginTop: 10 }}
-            >
-              {q.answers.map((a) => {
-                const isSelected = chosen.includes(a.id);
-                const locked = attempt.onlyOnce && chosen.length > 0 && !isMultiple;
-
-                return (
-                  <button
-                    key={a.id}
-                    disabled={locked && !isSelected}
-                    onClick={() => toggleAnswer(q, a.id)}
-                    style={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      gap: 10,
-                      width: '100%',
-                      textAlign: 'left' as const,
-                      padding: '11px 16px',
-                      borderRadius: 'var(--radius-sm)',
-                      border: isSelected
-                        ? '1.5px solid var(--color-primary)'
-                        : '1.5px solid rgba(234,88,12,0.20)',
-                      background: isSelected ? 'rgba(234,88,12,0.12)' : 'var(--color-bg)',
-                      color: isSelected ? 'var(--color-primary)' : 'var(--color-text)',
-                      cursor: locked && !isSelected ? 'not-allowed' : 'pointer',
-                      opacity: locked && !isSelected ? 0.5 : 1,
-                      fontSize: '0.9rem',
-                      fontWeight: isSelected ? 600 : 400,
-                      transition: 'all 0.15s',
-                    }}
-                    onMouseEnter={(e) => {
-                      if (!isSelected && !(locked && !isSelected)) {
-                        (e.currentTarget as HTMLButtonElement).style.background =
-                          'rgba(234,88,12,0.06)';
-                        (e.currentTarget as HTMLButtonElement).style.borderColor =
-                          'rgba(234,88,12,0.40)';
-                      }
-                    }}
-                    onMouseLeave={(e) => {
-                      if (!isSelected) {
-                        (e.currentTarget as HTMLButtonElement).style.background = 'var(--color-bg)';
-                        (e.currentTarget as HTMLButtonElement).style.borderColor =
-                          'rgba(234,88,12,0.20)';
-                      }
-                    }}
-                  >
-                    {/* Marcador: ☑ en MULTIPLE, ◉ en SINGLE/TRUE_FALSE */}
-                    <span
-                      aria-hidden
-                      style={{
-                        flexShrink: 0,
-                        width: 18,
-                        height: 18,
-                        borderRadius: isMultiple ? 4 : '50%',
-                        border: isSelected
-                          ? '2px solid var(--color-primary)'
-                          : '2px solid rgba(234,88,12,0.30)',
-                        background: isSelected ? 'var(--color-primary)' : 'transparent',
-                        display: 'inline-flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        color: '#fff',
-                        fontSize: '0.7rem',
-                        fontWeight: 900,
-                        lineHeight: 1,
-                      }}
-                    >
-                      {isSelected ? (isMultiple ? '✓' : '●') : ''}
-                    </span>
-                    <span>{a.text}</span>
-                  </button>
-                );
-              })}
-            </div>
-          </div>
-        );
-      })}
-
-      {/* Boton entregar */}
-      <div style={{ marginTop: 8, display: 'flex', justifyContent: 'flex-end' }}>
-        <button
-          className="btn btn-primary"
-          style={{
-            padding: '12px 28px',
-            fontSize: '1rem',
-            opacity: answeredCount < totalCount ? 0.45 : 1,
-            cursor: answeredCount < totalCount ? 'not-allowed' : 'pointer',
-          }}
-          disabled={answeredCount < totalCount || isLoading}
-          onClick={handleSubmit}
-        >
-          {isLoading ? 'Entregando...' : 'Entregar examen'}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// ─── Componente: Resultados ───────────────────────────────────────────────────
-
-function ResultsStep({
-  result,
-  scopeTitle,
-  courseId,
-  moduleId,
-  onRepeat,
-  onBack,
-  historyItems,
-}: {
-  result: ExamAttemptResult;
-  scopeTitle: string;
-  courseId?: string;
-  moduleId?: string;
-  onRepeat: () => void;
-  onBack: () => void;
-  historyItems: {
-    attemptId: string;
-    score: number | null;
-    numQuestions: number;
-    submittedAt: string | null;
-  }[];
-}) {
-  const { data: certs } = useMyCertificates();
-  const examCertType = courseId ? 'COURSE_EXAM' : 'MODULE_EXAM';
-  const examScopeId = courseId ?? moduleId;
-  const examCert = certs?.find((c) => c.scopeId === examScopeId && c.type === examCertType);
-
-  const passed = result.score >= 50;
-
-  return (
-    <div>
-      {/* Score grande */}
-      <div
-        style={{
-          background: 'var(--gradient-hero)',
-          borderRadius: 'var(--radius-xl)',
-          padding: '40px 36px',
-          textAlign: 'center' as const,
-          marginBottom: 24,
-          position: 'relative' as const,
-          overflow: 'hidden',
-        }}
-      >
-        {/* Glow decorativo */}
-        <div
-          style={{
-            position: 'absolute' as const,
-            top: -80,
-            right: -80,
-            width: 280,
-            height: 280,
-            background: `radial-gradient(circle, ${scoreColor(result.score)}22 0%, transparent 70%)`,
-            pointerEvents: 'none',
-          }}
-        />
-
-        <div
-          style={{
-            fontSize: '5rem',
-            fontWeight: 900,
-            background: scoreGradient(result.score),
-            WebkitBackgroundClip: 'text',
-            WebkitTextFillColor: 'transparent',
-            backgroundClip: 'text',
-            lineHeight: 1,
-            letterSpacing: '-0.03em',
-            marginBottom: 8,
-          }}
-        >
-          {result.score.toFixed(1)}%
-        </div>
-
-        <div style={{ fontSize: '1rem', color: 'rgba(255,255,255,0.65)', marginBottom: 16 }}>
-          {result.correctCount} de {result.numQuestions} respuestas correctas
-        </div>
-
-        {passed && (
-          <div
-            style={{
-              display: 'inline-flex',
-              alignItems: 'center',
-              gap: 8,
-              padding: '6px 18px',
-              borderRadius: 999,
-              background: 'rgba(22,163,74,0.15)',
-              border: '1px solid rgba(22,163,74,0.35)',
-              color: '#4ade80',
-              fontWeight: 700,
-              fontSize: '0.875rem',
-            }}
-          >
-            Certificado emitido
-          </div>
-        )}
-      </div>
-
-      {/* Correcciones */}
-      <div className="vkb-card" style={{ padding: '24px', marginBottom: 20 }}>
-        <h3
-          style={{
-            fontWeight: 700,
-            marginBottom: 16,
-            fontSize: '0.95rem',
-            color: 'var(--color-text)',
-          }}
-        >
-          Correcciones
-        </h3>
-        <div style={{ display: 'flex', flexDirection: 'column' as const, gap: 10 }}>
-          {result.corrections.map((c, i) => (
-            <div
-              key={c.questionId}
-              style={{
-                borderRadius: 'var(--radius-sm)',
-                padding: '14px 16px',
-                borderLeft: `4px solid ${c.isCorrect ? '#16a34a' : '#dc2626'}`,
-                background: c.isCorrect ? 'rgba(22,163,74,0.05)' : 'rgba(220,38,38,0.04)',
-                border: `1px solid ${c.isCorrect ? 'rgba(22,163,74,0.20)' : 'rgba(220,38,38,0.18)'}`,
-                borderLeftWidth: 4,
-              }}
-            >
-              <div
-                style={{
-                  fontWeight: 600,
-                  marginBottom: 6,
-                  fontSize: '0.9rem',
-                  color: 'var(--color-text)',
-                  lineHeight: 1.4,
-                }}
-              >
-                <span style={{ marginRight: 6, fontSize: '0.85rem' }}>
-                  {c.isCorrect ? '✓' : '✗'}
-                </span>
-                {i + 1}. {c.questionText}
-              </div>
-
-              {(() => {
-                // Para MULTIPLE muestra todas las seleccionadas/correctas; para
-                // SINGLE/TRUE_FALSE cae en los campos legacy con un único valor.
-                const selectedTexts =
-                  c.selectedAnswerTexts && c.selectedAnswerTexts.length > 0
-                    ? c.selectedAnswerTexts
-                    : c.selectedAnswerText
-                      ? [c.selectedAnswerText]
-                      : [];
-                const correctTexts =
-                  c.correctAnswerTexts && c.correctAnswerTexts.length > 0
-                    ? c.correctAnswerTexts
-                    : c.correctAnswerText
-                      ? [c.correctAnswerText]
-                      : [];
-                return (
-                  <>
-                    {selectedTexts.length > 0 ? (
-                      <div
-                        style={{
-                          fontSize: '0.82rem',
-                          color: c.isCorrect ? 'var(--color-text-muted)' : '#dc2626',
-                          marginBottom: c.isCorrect ? 0 : 4,
-                        }}
-                      >
-                        {selectedTexts.length === 1 ? 'Tu respuesta' : 'Tus respuestas'}:{' '}
-                        {selectedTexts.join(' · ')}
-                      </div>
-                    ) : (
-                      <div style={{ fontSize: '0.82rem', color: 'var(--color-text-muted)' }}>
-                        Sin respuesta
-                      </div>
-                    )}
-
-                    {!c.isCorrect && correctTexts.length > 0 && (
-                      <div
-                        style={{
-                          fontSize: '0.82rem',
-                          color: '#16a34a',
-                          fontWeight: 600,
-                          marginTop: 4,
-                        }}
-                      >
-                        {correctTexts.length === 1 ? 'Respuesta correcta' : 'Respuestas correctas'}:{' '}
-                        {correctTexts.join(' · ')}
-                      </div>
-                    )}
-                  </>
-                );
-              })()}
-            </div>
-          ))}
-        </div>
-      </div>
-
-      {/* Historial */}
-      {historyItems.filter((h) => h.submittedAt).length > 1 && (
-        <div className="vkb-card" style={{ padding: '24px', marginBottom: 24 }}>
-          <h4
-            style={{
-              fontWeight: 700,
-              marginBottom: 14,
-              fontSize: '0.9rem',
-              color: 'var(--color-text)',
-            }}
-          >
-            Historial de intentos
-          </h4>
-          {historyItems
-            .filter((h) => h.submittedAt)
-            .map((h) => (
-              <div
-                key={h.attemptId}
-                style={{
-                  display: 'flex',
-                  justifyContent: 'space-between',
-                  alignItems: 'center',
-                  padding: '9px 0',
-                  borderBottom: '1px solid var(--color-border)',
-                  fontSize: '0.875rem',
-                  gap: 12,
-                }}
-              >
-                <span style={{ color: 'var(--color-text-muted)' }}>
-                  {new Date(h.submittedAt!).toLocaleDateString('es-ES')}
-                </span>
-                <span style={{ color: 'var(--color-text-muted)' }}>{h.numQuestions} preguntas</span>
-                <span
-                  style={{
-                    fontWeight: 800,
-                    color: scoreColor(h.score ?? 0),
-                  }}
-                >
-                  {h.score?.toFixed(1)}%
-                </span>
-              </div>
-            ))}
-        </div>
-      )}
-
-      {/* Acciones */}
-      <div style={{ display: 'flex', gap: 12, flexWrap: 'wrap' as const }}>
-        <button className="btn btn-ghost" onClick={onBack}>
-          Volver al curso
-        </button>
-        <button className="btn btn-primary" onClick={onRepeat}>
-          Repetir examen
-        </button>
-        <button className="btn btn-ghost" onClick={() => downloadExamPdf(result, scopeTitle)}>
-          Descargar PDF examen
-        </button>
-        {examCert && (
-          <button
-            className="btn"
-            style={{
-              background: 'linear-gradient(135deg, #16a34a 0%, #22c55e 100%)',
-              color: '#fff',
-              boxShadow: '0 4px 16px rgba(22,163,74,0.30)',
-            }}
-            onClick={() => downloadCertificatePdf(examCert)}
-          >
-            Descargar certificado
-          </button>
-        )}
-      </div>
-    </div>
-  );
-}
 
 // ─── Componente principal ─────────────────────────────────────────────────────
